% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypes.R
\name{filter_gtype}
\alias{filter_gtype}
\title{Filter genotypes from experiment for the analysis}
\usage{
filter_gtype(gtypemap, plantprefix, rejectgtypes = F)
}
\arguments{
\item{gtypemap}{dataframe of the genotypemap}

\item{plantprefix}{prefix for filename that include tray and roi #s to be included. 'all' is preserved for no filtering.}

\item{rejectgtypes}{logical. default=FALSE. TRUE means the complementary set of genotypes specified by plantprefix file will be analyzed.}
}
\description{
Filter genotypes from experiment for the analysis
}
\details{
this should be run after \code{get_genotypemap()}
the plantprefix is part of \verb{<plantprefix>_gtypes_[A-Z].csv}. Each such file with a common plantprefix identifies the poster child individuals that get used for a collage. The suffix A,B,... are to allow multiple columns of genotypes in the collage. All suffixed files are read in this function.
}
