% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypes.R
\name{genotype_colors}
\alias{genotype_colors}
\title{Define standard colors for genotypes}
\usage{
genotype_colors(gtypemap, colorpal = NULL)
}
\arguments{
\item{gtypemap}{dataframe of the genotypemap}

\item{colorpal}{optional color palette to use for the genotypes.}
}
\value{
named vector of colors with genotypes as names
}
\description{
Define standard colors for genotypes
}
\details{
This should be run after \code{get_genotypemap()} which takes filename for genotypemap and creates and orders gtype levels
Each set of genotypes are to be given a custom color with wild-type 'black'. The output of this function is designed to be used with scale_<color/fill>_manual values argument. The colors of the genotypes will be consistent even if only a subset of genotypes are plotted.
The default colors are 'black' for wild-type and then RColorBrewer palettes Set1 and Pastel1
}
