% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypes.R
\name{get_genotypemap}
\alias{get_genotypemap}
\title{Get and format genotype map of an experiment}
\usage{
get_genotypemap(fn, wtcode = "WT", gtypeorder = NULL)
}
\arguments{
\item{fn}{filename of genotypemap usually stored at data/genotype_map.csv in the experiment folder}

\item{wtcode}{string used to identify wild-type}

\item{gtypeorder}{vector of genotype levels in custom order}
}
\value{
tibble, with gtype as a factor.
}
\description{
Get and format genotype map of an experiment
}
\details{
The standard columns for this file are plantbarcode, roi, gtype and should map each unique position in the hotel to a genotype
the genotypemap is part of every experiment and maps the barcode and roi # with the genotype.
}
