% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{italic_labels}
\alias{italic_labels}
\title{Format plot labels with italics for mutant names}
\usage{
italic_labels(names)
}
\arguments{
\item{names}{character vector of names}
}
\description{
Format plot labels with italics for mutant names
}
\details{
designed to be used in scale_\emph{_} for the labels argument. wild-type can be 'WT' or 'Col-0'
}
