% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_waterdata.R
\name{query_waterdata}
\alias{query_waterdata}
\title{Fetch watering data from LT database}
\usage{
query_waterdata(config, expername, jobname)
}
\arguments{
\item{config}{configuration for database as json file}

\item{expername}{character or vector of characters; experiment name(s) in LT database}

\item{jobname}{character or vector of characters; job name(s) (internally stored as \code{comment}) in LT database}
}
\description{
Fetch watering data from LT database
}
\details{
The \% character is supported as a multicharacter wildcard and
_ is supported as a single character wildcard for experiment and job names.
names are case-sensitive!
}
