% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_change.R
\name{reference_phenotype}
\alias{reference_phenotype}
\title{Isolate a reference phenotype}
\usage{
reference_phenotype(
  dF,
  statcol_central = "med",
  statcol_spread = "sd",
  refcol = "gtype",
  refvalue = "WT",
  ...
)
}
\arguments{
\item{refcol}{column name}

\item{refvalue}{string to identify reference}

\item{...}{additional columns to keep in the outpur}
}
\value{
dataframe
}
\description{
Isolate a reference phenotype
}
