% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_phenotypes.R
\name{summarise_phenotypes}
\alias{summarise_phenotypes}
\title{Summarise phenotypes}
\usage{
summarise_phenotypes(
  dF = NULL,
  phenotype = NULL,
  addgrp = NULL,
  statistic = list(med = median, sd = sd, se = se),
  na.rm = TRUE
)
}
\arguments{
\item{dF}{dataframe of phenotypes for every individual}

\item{phenotype}{column name or vector of column names of phenotypes to summarise. unquoted.}

\item{addgrp}{additional columns to group by. column name or vector of column names}

\item{statistic}{named list of function(s) to summarise with. name of your list entry will be used to name the new column}

\item{na.rm}{logical. remove rows with NA. default is TRUE.}
}
\value{
dataframe with summary columns named as \verb{\{statistic\}_\{phenotype\}}
}
\description{
Summarise phenotypes
}
\details{
the default is to group by \code{gtype}, \code{idate}, \code{jobdate}. For PSII data you need to additionally specify \code{parameter}
}
\examples{
summarise_phenotypes(mtcars,phenotype = c(hp, disp), statistic = list(med=median), na.rm=T)
}
