% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{theme_darkcowplot}
\alias{theme_darkcowplot}
\title{Dark cowplot theme}
\usage{
theme_darkcowplot(bkgrdcol = "grey40", ...)
}
\arguments{
\item{bkgrdcol}{valid color for background. default = 'grey40'}

\item{...}{arguments for \code{theme_cowplot()}}
}
\description{
Dark cowplot theme
}
\details{
see \code{theme_cowplot()} for input arguments.
panel.background is set to grey40 to better accommodate black and light colors when number of genotypes is >9
legend.key is also set to grey40
}
